Colocate scores export snapshot gathering with payload resolution.
Move gather_scores_inference_snapshot and its helpers into export_materialization so exports.py stays catalog wiring and ensure orchestration.
